requirements.txt
Keras==2.0.4
pymongo==3.3.0
Flask_RESTful==0.3.6
pandas==0.19.2
requests==2.10.0
Theano==0.9.0
names==0.3.0
Flask==0.10.1
numpy==1.12.1
Pillow==4.1.1
boto==2.47.0
beautifulsoup4==4.6.0
